fix: review findings — Scrub panic path, janitor cleanup race, arena stale handle, uint64 keys, docs (#93)
Fixes from the adversarial review of the core module. Every behavioural
fix
carries a regression test that was run against the unfixed code first
and
shown to fail; the forced interleavings use two cold-path test seams
(`janitorWipeTestHook`, `rLockSlowTestHook`) rather than racing for
them.

## Fixed

- **`Scrub`/`ScrubErr` (legacy path) did not scrub live frames on the
panic
path.** The deferred wipe ran from `gopanic`'s frame, *below* the frames
still live at the panic. Measured before the fix: 2048/2048 marker bytes
survived, same as no `Scrub`. The window now runs `fn` through a helper
that recovers the panic (`scrubCall`, mirroring
`runtime/secret.doHelper`),
  clears the vector file and burns the band once those frames are dead,
then re-raises from `Scrub`. `runtime.Goexit` is documented as
best-effort
  (deferred backstop only). The vecclear control was reshaped to match;
  `TestNoHeapEscape_Scrub` stays at 0 allocs. Test:
  `TestScrub_ScrubsLiveFramesOnPanic` (+ `ScrubErr` twin).

- **GC cleanup racing `WipeAllSecrets` stranded a locked mapping.** The
cleanup consulted the registry before locking; the passes held the
region
in *neither* set for the length of the wipe. Two changes: the passes
move
  the region between sets in one registry critical section
  (`moveToWipedIf`), and `release(key, false)` now peeks → locks →
  `takeAnyIf`, the passes' own order. Test:
`TestJanitorRelease_ConcurrentWithWipePass_ReclaimsMapping` (both
passes;
failed before with "release returned early, mapping still registered").

- **`ArenaSlot.WithBytesErr` generation check was check-then-use.** It
ran
before `rLock`; a borrower parked behind a queued writer resumed with a
  slice into a slot that had been released, re-acquired and written. The
  check now runs under `rLock` immediately before the slice is produced;
  `alloc` is still not taken on the borrow path. Test:
  `TestArenaSlot_StaleHandleRefusedUnderLock` (failed before: the stale
  callback ran and read the next tenant's bytes).

- **Janitor keys are `uint64` end to end** (maps, owner fields, cleanup
argument; `LockOrder` already returned `uint64`). `register` refuses a
key
already in either set, or zero, with an error the constructors propagate
  after wiping and freeing the region. Tests:
  `TestJanitorRegister_RefusesKeyCollision` / `_RefusesZeroKey`.
  `GOARCH=386 go vet` and `go test -c` pass.

- **`Probe` no longer fires the "INSECURE fallback in use" warning** on
stub
  platforms; the warning moved to `gateInsecure` and fires only when
  `WithInsecureFallback` is used.

## Docs

- `securebuf.go` header: lists every exclusive-lock taker; stale
  `raw`/"Nil out data and raw" comments fixed.
- `DESIGN.md`/`PITFALLS.md`: the GC is non-moving for the heap; the real
  off-heap rationale (no lock/guard/protect on heap pages, no zeroing on
  free, stacks do move, unwiped reuse). Cache-flush rationale corrected:
  caches are coherent, the flush only shortens DRAM remanence (DMA /
  cold-boot — out of scope; defence in depth). amd64 wipe described as
  `REP STOSB` + fences.
- `doc.go`: wipe is assembly only on amd64/arm64; Windows "no swap"
caveat.
- Windows `VirtualLock` overstatement: README cell → `⚠ VirtualLock ‡`
with
footnote (no other rows touched), `Capabilities.Mlocked` doc, and a new
  `Warnings()` line on Windows. `harden_windows.go` documents why the
  working-set minimum stays soft (`HARDWS_MIN_DISABLE`): trimming never
touches locked pages, so a hard minimum buys the secrets nothing;
whether
it exempts a process from whole-working-set outswap is undocumented and
is
  stated as not measured.
- `slotMeta.generation` comment matches the new check placement.

## Verification (Windows/amd64, GOWORK=off)

`gofmt -l` clean · `go vet -unsafeptr=false ./...` clean (plain `go vet`
shows only the four pre-existing `unsafeptr` findings) · `go test
-count=1
./...` green · `go test -race -count=1 .` green · `GOARCH=386 go vet`
and
`go test -c` OK.


---------

+++ securebuf_lockorder.go
@@ -12,5 +12,5 @@ func (b *SecureBuffer) LockOrder() uint64 {
 	if b == nil {
 		return 0
 	}
-	return uint64(b.janitorKey)
+	return b.janitorKey
 }
